requirements.txt
asgiref==3.4.1
Django==3.2.6
pytz==2021.1
sqlparse==0.4.1
paho-mqtt==1.5.1
